Modularize bandwidth parser and add regression seed
