+++ crates/bandwidth/proptest-regressions/parse/tests.txt
@@ -1,0 +1,7 @@
+# Seeds for failure cases proptest has generated in the past. It is
+# automatically read and these particular cases re-run before any
+# novel cases are generated.
+#
+# It is recommended to check this file in to source control so that
+# everyone who runs the test benefits from these saved cases.
+cc 2656a422f3a1f888a98cdb87bf082e1c645c449512fefead95e4dd98673f1903 # shrinks to value = 1
